Issue #10981: Moved VsiteCurrentFilterTest to drupal-test-traits
