+++ profile/modules/vsite/tests/modules/vsite_module_test/vsite_module_test.info.yml
@@ -3,4 +3,6 @@ name: Vsite Module Test
 description: Testing module for token functionality.
 package: Testing
 core: 8.x
+dependencies:
+  - views
 hidden: TRUE
