+++ profile/modules/vsite/tests/src/ExistingSite/VsiteCurrentFilterTest.php
@@ -1,16 +1,10 @@
 <?php
 
-namespace Drupal\Tests\vsite\Kernel;
+namespace Drupal\Tests\vsite\ExistingSite;
 
-use Drupal\Core\DependencyInjection\ContainerBuilder;
 use Drupal\group\Entity\Group;
-use Drupal\node\Entity\Node;
-use Drupal\node\Entity\NodeType;
-use Drupal\Tests\user\Traits\UserCreationTrait;
-use Drupal\Tests\views\Kernel\ViewsKernelTestBase;
-use Drupal\views\Tests\ViewTestData;
 use Drupal\views\Views;
-use PHPUnit\Framework\Error\Deprecated;
+use weitzman\DrupalTestTraits\ExistingSiteBase;
 
 /**
  * Test the Current Vsite Views Filter.
@@ -20,28 +14,7 @@ use PHPUnit\Framework\Error\Deprecated;
  * @group kernel
  * @coversDefaultClass \Drupal\vsite\Plugin\views\filter\VsiteCurrentFilter
  */
-class VsiteCurrentFilterTest extends ViewsKernelTestBase {
-
-  use UserCreationTrait;
-
-  /**
-   * Modules to enable.
-   *
-   * @var array
-   */
-  public static $modules = [
-    'system',
-    'field',
-    'node',
-    'group',
-    'gnode',
-    'views',
-    'text',
-    'filter',
-    'purl',
-    'vsite',
-    'vsite_module_test',
-  ];
+class VsiteCurrentFilterTest extends ExistingSiteBase {
 
   /**
    * Views used by this test.
@@ -82,27 +55,14 @@ class VsiteCurrentFilterTest extends ViewsKernelTestBase {
   /**
    * {@inheritdoc}
    */
-  protected function setUp($import_test_views = TRUE) {
-    Deprecated::$enabled = FALSE;
-    parent::setUp(FALSE);
-    $this->installSchema('node', 'node_access');
-    $this->installEntitySchema('user');
-    $this->installEntitySchema('node');
-    $this->installEntitySchema('group_type');
-    $this->installEntitySchema('group');
-    $this->installEntitySchema('group_content_type');
-    $this->installEntitySchema('group_content');
-
-    $this->installConfig(['field', 'node', 'group', 'vsite']);
+  public function setUp($import_test_views = TRUE) {
+    parent::setUp();
 
     // Set the current user so group creation can rely on it.
     $this->container->get('current_user')->setAccount($this->createUser());
 
     /** @var \Drupal\Core\Entity\EntityTypeManagerInterface $entityTypeManager */
     $entityTypeManager = $this->container->get('entity_type.manager');
-
-    $node_type = NodeType::create(['type' => 'page', 'name' => t('Page')]);
-    $node_type->save();
 
     /** @var \Drupal\group\Entity\GroupTypeInterface $group_type */
     $group_type = $entityTypeManager->getStorage('group_type')->load('personal');
@@ -112,8 +72,6 @@ class VsiteCurrentFilterTest extends ViewsKernelTestBase {
     $storage = $entityTypeManager->getStorage('group_content_type');
     $plugin = $storage->createFromPlugin($group_type, 'group_node:page');
     $plugin->save();
-
-    ViewTestData::createTestViews(get_class($this), ['vsite_module_test']);
 
     $this->group = Group::create([
       'type' => 'personal',
@@ -128,19 +86,19 @@ class VsiteCurrentFilterTest extends ViewsKernelTestBase {
     $otherGroup->save();
 
     // Create the nodes we'll be displaying (or not) in the view.
-    $this->ungroupedNode = Node::create([
+    $this->ungroupedNode = $this->createNode([
       'type' => 'page',
       'title' => 'Ungrouped',
     ]);
     $this->ungroupedNode->save();
 
-    $this->groupedNode = Node::create([
+    $this->groupedNode = $this->createNode([
       'type' => 'page',
       'title' => 'Grouped',
     ]);
     $this->groupedNode->save();
 
-    $otherNode = Node::create([
+    $otherNode = $this->createNode([
       'type' => 'page',
       'title' => 'OtherGroup',
     ]);
@@ -150,39 +108,6 @@ class VsiteCurrentFilterTest extends ViewsKernelTestBase {
     $otherGroup->addContent($otherNode, $plugin->getContentPluginId());
 
     $this->vsiteContextManager = $this->container->get('vsite.context_manager');
-  }
-
-  /**
-   * Sets up the configuration and schema of views and views_test_data modules.
-   *
-   * Because the schema of views_test_data.module is dependent on the test
-   * using it, it cannot be enabled normally.
-   */
-  protected function setUpFixtures() {
-    // First install the system module. Many Views have Page displays have menu
-    // links, and for those to work, the system menus must already be present.
-    $this->installConfig(['system']);
-
-    /** @var \Drupal\Core\State\StateInterface $state */
-    $state = $this->container->get('state');
-
-    \Drupal::service('plugin.manager.views.filter')->clearCachedDefinitions();
-
-    $this->installConfig(['views', 'vsite_module_test']);
-
-    $this->container->get('router.builder')->rebuild();
-  }
-
-  /**
-   * {@inheritdoc}
-   *
-   * Set a mock for the purl processor so we don't have to install purl.
-   */
-  public function register(ContainerBuilder $container) {
-    $purlPathProcessor = $this->createMock('\Drupal\purl\PathProcessor\PurlContextOutboundPathProcessor');
-    $container->set('purl.outbound_path_processor', $purlPathProcessor);
-
-    parent::register($container);
   }
 
   /**
